fix: add extra_alert_labels for injected generic alerts
