+++ tests/unit/test_alert_labels.py
@@ -178,9 +178,9 @@ def test_extra_alerts_config():
 
     for group in alert_rules["groups"]:
         for rule in group["rules"]:
+            assert rule["labels"]["environment"] == "PRODUCTION"
+            assert rule["labels"]["zone"] == "Mars"
             if "grafana_agent_alertgroup_alerts" in group["name"]:
-                assert rule["labels"]["environment"] == "PRODUCTION"
-                assert rule["labels"]["zone"] == "Mars"
                 assert (
                     rule["labels"]["juju_application"] == "primary"
                     or rule["labels"]["juju_application"] == "subordinate"
